chore: ignore unmaintained warning for `net2` in `cargo audit`
`cargo audit` currently emits a warning that the `net2` crate is
unmaintained. We only depend on `net2` as a transitive dependency of
older versions of `mio`, which we depend on via `tokio` 0.1.
`tracing-futures` has feature flags for supporting `tokio` 0.1, which we
can't remove until the next breaking change. `tokio` 0.1 won't be
updated, so as long as `tracing-futures` supports tokio 0.1, we can't
really get rid of the `net2` dependency.

Therefore, this commit adds a `.cargo/audit.toml` to just ignore the
warning. It only effects users who are using compatibility features for
*other* unmaintained libraries, anyway. Eventually, when we drop `tokio`
0.1 support entirely, we can remove the `ignore` for this warning.