subscriber: `<None as Layer>::event_enabled` should be `true` (#2193)
## Motivation

This is wrong.

## Solution

Make it unwrong.

As described by the documentation on `Layer::event_enabled`, the
return value sets the global filtering of an event. This code used to
say that that `Option::<impl Layer>::None` existing in the
layer stack should disable any layer in the stack seeing
events